+++ db/migrate/20160602202653_create_instagram_photos.rb
@@ -1,0 +1,12 @@
+class CreateInstagramPhotos < ActiveRecord::Migration
+  def change
+    create_table :instagram_photos do |t|
+      t.string     :short_code, null: false
+      t.references :trail,      null: false
+      t.boolean    :processed,  null: false, default: false
+      t.text       :json
+
+      t.timestamps null: false
+    end
+  end
+end